Refactor: move new annotation classes to dedicated module
